Finished filing cabinets
Finished a filling cabinet prefab that has drawers that can be opened, as well as a variant that is just flat with no openable drawers. The one with drawers doesn't have faces inside the frame so the drawers should always remain 'attached' to the frame and not loose on their own.